move ArtifactAppDeployer class to internal.io.deployment package
